+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 - Support for updating `license_model` for `oci_autonomous_database` resource
 - Support for updating `static_routes` and new `cpe_local_identifier` in `oci_core_ipsec` resource for improved VPN service usability
 - Support for updating `whitelisted_ips` in `autonomous_database`. Note: Cannot be used during creation.
-- Support tagging for Dynamic Groups in Identity 
+- Support tagging for Dynamic Groups in Identity
+- Singular data sources will reuse resource schema 
 
 ## 3.22.0 (April 10, 2019)